+++ includes/plugins/wc-memberships/class-memberships.php
@@ -16,7 +16,8 @@ defined( 'ABSPATH' ) || exit;
  */
 class Memberships {
 
-	const GATE_CPT = 'np_memberships_gate';
+	const GATE_CPT  = 'np_memberships_gate';
+	const CRON_HOOK = 'np_memberships_fix_expired_memberships';
 
 	/**
 	 * Whether the gate has been rendered in this execution.
@@ -63,6 +64,10 @@ class Memberships {
 		add_filter( 'newspack_gate_content', 'wp_filter_content_tags' );
 		add_filter( 'newspack_gate_content', 'wp_replace_insecure_home_url' );
 		add_filter( 'newspack_gate_content', 'do_shortcode', 11 ); // AFTER wpautop().
+
+		// Scheduled fix for prematurely expired memberships.
+		add_action( 'init', [ __CLASS__, 'cron_init' ] );
+		add_action( self::CRON_HOOK, [ __CLASS__, 'fix_expired_memberships_for_active_subscriptions' ] );
 
 		include __DIR__ . '/class-block-patterns.php';
 		include __DIR__ . '/class-metering.php';
@@ -868,6 +873,89 @@ class Memberships {
 		}
 
 		return $has_access;
+	}
+
+	/**
+	 * Deactivate the cron job.
+	 */
+	public static function cron_deactivate() {
+		\wp_clear_scheduled_hook( self::CRON_HOOK );
+	}
+
+	/**
+	 * Schedule an hourly cron job to check for and fix expired memberships linked to active subscriptions.
+	 */
+	public static function cron_init() {
+		\register_deactivation_hook( NEWSPACK_PLUGIN_FILE, [ __CLASS__, 'cron_deactivate' ] );
+
+		if ( ! wp_next_scheduled( self::CRON_HOOK ) ) {
+			\wp_schedule_event( time(), 'hourly', self::CRON_HOOK );
+		}
+	}
+
+	/**
+	 * Ensure that memberships tied to active subscriptions haven't expired prematurely.
+	 * Will loop through all active subscriptions on the site and check all memberships associated with it.
+	 */
+	public static function fix_expired_memberships_for_active_subscriptions() {
+		if ( ! function_exists( 'wc_memberships' ) ) {
+			return;
+		}
+
+
+		$max_per_batch           = 100;
+		$processed_batches       = 0;
+		$reactivated_memberships = 0;
+		$active_subscriptions    = WooCommerce_Connection::get_batch_of_active_subscriptions( $max_per_batch, $processed_batches );
+		if ( empty( $active_subscriptions ) ) {
+			return;
+		}
+
+		Logger::log( __( 'Checking for expired memberships linked to active subscriptions...', 'newspack-plugin' ) );
+
+		$active_membership_statuses = \wc_memberships()->get_user_memberships_instance()->get_active_access_membership_statuses();
+		while ( ! empty( $active_subscriptions ) ) {
+			$subscription = array_shift( $active_subscriptions );
+			try {
+				$memberships = \wc_memberships_get_memberships_from_subscription( $subscription );
+				if ( $memberships ) {
+					foreach ( $memberships as $membership ) {
+						// If the membership is not active and has an end date in the past, reactivate it.
+						if ( $membership && ! $membership->has_status( $active_membership_statuses ) && $membership->has_end_date() && $membership->get_end_date( 'timestamp' ) < time() ) {
+							$membership->set_end_date(); // Clear the end date.
+							$membership->update_status( 'active' ); // Reactivate the membership.
+							$reactivated_memberships++;
+						}
+					}
+				}
+			} catch ( Exception $e ) {
+				// Log the error.
+				Logger::log(
+					sprintf(
+						// Translators: %1$d is the membership ID, %2$s is the subscription ID.
+						__( 'Failed to reactivate membership %1$d linked to active subscription %2$s.', 'newspack-plugin' ),
+						$membership->get_id(),
+						$subscription->get_id()
+					)
+				);
+			}
+
+			// Get the next batch of active subscriptions.
+			if ( empty( $active_subscriptions ) ) {
+				$processed_batches++;
+				$active_subscriptions = WooCommerce_Connection::get_batch_of_active_subscriptions( $max_per_batch, $processed_batches * $max_per_batch );
+			}
+		}
+
+		if ( 0 < $reactivated_memberships ) {
+			Logger::log(
+				sprintf(
+					// Translators: %d is the number of reactivated memberships.
+					__( 'Reactivated %d memberships linked to active subscriptions.', 'newspack-plugin' ),
+					$reactivated_memberships
+				)
+			);
+		}
 	}
 }
 Memberships::init();